+++ src/Authorization/Serializer/EntityNormalizer.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Dbp\Relay\CoreBundle\Authorization\Serializer;
 
 use ApiPlatform\State\Pagination\PartialPaginatorInterface;
+use Dbp\Relay\CoreBundle\Authorization\AbstractAuthorizationService;
 use Dbp\Relay\CoreBundle\Helpers\Tools;
 use Symfony\Component\Serializer\Normalizer\NormalizerAwareInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerAwareTrait;
@@ -24,7 +25,8 @@ class EntityNormalizer implements NormalizerAwareInterface, NormalizerInterface
     private array $getOutputGroupsToAddForEntityInstanceCallbacks = [];
 
     /**
-     * for testing purposes.
+     * For testing purposes.
+     * NOTE: Must be reset manually, otherwise we lose the registered callbacks on Symfony's pre-request reset.
      */
     public function reset(): void
     {
@@ -33,27 +35,67 @@ class EntityNormalizer implements NormalizerAwareInterface, NormalizerInterface
     }
 
     /**
-     * @param callable(string): array $getOutputGroupsToAddCallback
+     * Adds output groups for the given entity class if the given condition is met.
+     *
+     * @param ?callable(string): bool $condition Condition for the groups to be added. If null, the groups are always added.
      */
-    public function registerGetOutputGroupsToAddForEntityClassCallback(string $entityClass,
-        callable $getOutputGroupsToAddCallback): void
+    public function showOutputGroupsForEntityClassIf(string $entityClass, array $outputGroups, ?callable $condition = null): void
     {
-        Tools::pushToSubarray($this->getOutputGroupsToAddForEntityClassCallbacks,
-            $entityClass, $getOutputGroupsToAddCallback);
+        $this->registerGetOutputGroupsToAddForEntityClassCallback($entityClass,
+            function (string $entityClass) use ($condition, $outputGroups) {
+                return $condition === null || $condition($entityClass) ? $outputGroups : [];
+            });
+    }
+
+    /**
+     * Adds output groups for the given entity class if the given roles are granted.
+     */
+    public function showOutputGroupsForEntityClassIfGrantedRoles(string $entityClass, array $outputGroups,
+        array $requiredRoles, AbstractAuthorizationService $authorizationService): void
+    {
+        $this->registerGetOutputGroupsToAddForEntityClassCallback($entityClass,
+            function () use ($requiredRoles, $outputGroups, $authorizationService) {
+                foreach ($requiredRoles as $requiredRole) {
+                    if (!$authorizationService->isGrantedRole($requiredRole)) {
+                        return [];
+                    }
+                }
+
+                return $outputGroups;
+            });
+    }
+
+    /**
+     * Adds output groups for the given entity instance if the given condition is met.
+     *
+     * @param callable(object, string): bool $condition
+     */
+    public function showOutputGroupsForEntityInstanceIf(string $entityClass, array $outputGroups, callable $condition): void
+    {
+        $this->registerGetOutputGroupsToAddForEntityInstanceCallback($entityClass,
+            function (object $entityInstance, string $entityClass) use ($condition, $outputGroups) {
+                return $condition($entityInstance, $entityClass) ? $outputGroups : [];
+            });
+    }
+
+    public function showOutputGroupsForEntityInstanceIfGrantedResourcePermissions(string $entityClass, array $outputGroups,
+        array $requiredResourcePermissions, AbstractAuthorizationService $authorizationService): void
+    {
+        $this->registerGetOutputGroupsToAddForEntityInstanceCallback($entityClass,
+            function (object $entityInstance) use ($requiredResourcePermissions, $outputGroups, $authorizationService) {
+                foreach ($requiredResourcePermissions as $requiredResourcePermission) {
+                    if (!$authorizationService->isGrantedResourcePermission($requiredResourcePermission, $entityInstance)) {
+                        return [];
+                    }
+                }
+
+                return $outputGroups;
+            });
     }
 
     public function tryGetGetOutputGroupsToAddForEntityClassCallback(string $entityClass): ?array
     {
         return $this->getOutputGroupsToAddForEntityClassCallbacks[$entityClass] ?? null;
-    }
-
-    /**
-     * @param callable(object, string): array $getOutputGroupsToAddCallback
-     */
-    public function registerGetOutputGroupsToAddForEntityInstanceCallback(string $entityClass, callable $getOutputGroupsToAddCallback): void
-    {
-        Tools::pushToSubarray($this->getOutputGroupsToAddForEntityInstanceCallbacks,
-            $entityClass, $getOutputGroupsToAddCallback);
     }
 
     public function normalize(mixed $object, ?string $format = null, array $context = []): string|float|int|bool|\ArrayObject|array|null
@@ -80,6 +122,25 @@ class EntityNormalizer implements NormalizerAwareInterface, NormalizerInterface
             && false === $data instanceof PartialPaginatorInterface
             && isset($this->getOutputGroupsToAddForEntityInstanceCallbacks[$context[self::RESOURCE_CLASS_CONTEXT_KEY] ?? null])
             && false === $this->wasAlreadyNormalized($data, $context);
+    }
+
+    /**
+     * @param callable(string): array $getOutputGroupsToAddCallback
+     */
+    protected function registerGetOutputGroupsToAddForEntityClassCallback(string $entityClass,
+        callable $getOutputGroupsToAddCallback): void
+    {
+        Tools::pushToSubarray($this->getOutputGroupsToAddForEntityClassCallbacks,
+            $entityClass, $getOutputGroupsToAddCallback);
+    }
+
+    /**
+     * @param callable(object, string): array $getOutputGroupsToAddCallback
+     */
+    protected function registerGetOutputGroupsToAddForEntityInstanceCallback(string $entityClass, callable $getOutputGroupsToAddCallback): void
+    {
+        Tools::pushToSubarray($this->getOutputGroupsToAddForEntityInstanceCallbacks,
+            $entityClass, $getOutputGroupsToAddCallback);
     }
 
     private function setAlreadyNormalized(object $entity, array &$context): void