+++ src/TestUtils/TestAuthorizationService.php
@@ -5,28 +5,15 @@ declare(strict_types=1);
 namespace Dbp\Relay\CoreBundle\TestUtils;
 
 use Dbp\Relay\CoreBundle\Authorization\AbstractAuthorizationService;
-use Dbp\Relay\CoreBundle\Authorization\Serializer\EntityNormalizer;
 use Dbp\Relay\CoreBundle\User\UserAttributeMuxer;
 use Dbp\Relay\CoreBundle\User\UserAttributeProviderProvider;
 use Dbp\Relay\CoreBundle\User\UserAttributeService;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\EventDispatcher\EventDispatcher;
 
-class TestAuthorizationService extends AbstractAuthorizationService
+class TestAuthorizationService
 {
     public const TEST_USER_IDENTIFIER = 'testuser';
     public const ADMIN_USER_IDENTIFIER = 'admin';
-
-    public static function create(?string $currentUserIdentifier = self::TEST_USER_IDENTIFIER,
-        array $currentUserAttributes = [], array $symfonyUerRoles = [],
-        bool $isAuthenticated = true, bool $isServiceAccount = false): TestAuthorizationService
-    {
-        $testAuthorizationService = new TestAuthorizationService();
-        self::setUp($testAuthorizationService, $currentUserIdentifier, $currentUserAttributes, $symfonyUerRoles,
-            $isAuthenticated, $isServiceAccount);
-
-        return $testAuthorizationService;
-    }
 
     public static function setUp(AbstractAuthorizationService $authorizationService,
         ?string $currentUserIdentifier = self::TEST_USER_IDENTIFIER, array $currentUserAttributes = [], array $symfonyUserRoles = [],
@@ -41,13 +28,14 @@ class TestAuthorizationService extends AbstractAuthorizationService
             new UserAttributeMuxer(new UserAttributeProviderProvider([$userAttributeProvider]), new EventDispatcher()));
 
         $authorizationService->__injectUserAttributeService($userAttributeService);
-        $authorizationService->__injectEntityNormalizer(new EntityNormalizer());
     }
 
-    public static function clearAuthorizationServiceRequestCaches(ContainerInterface $container, string $authorizationServiceId): void
+    /**
+     * Resets the internal state (e.g. request caches).
+     * Should be called between requests when performing multiple requests in a single test case.
+     */
+    public static function reset(AbstractAuthorizationService $authorizationService): void
     {
-        /** @var AbstractAuthorizationService $authorizationService */
-        $authorizationService = $container->get($authorizationServiceId);
-        $authorizationService->clearRequestCaches();
+        $authorizationService->reset();
     }
 }